pyproject.toml
[project]
name = "instalily-onsite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40",
    "sqlalchemy>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[tool.uv.workspace]
members = [
    "backend",
]
